Renamed articles tables component
