+++ app/frontend/components/articles-table.component.tsx
