+++ app/frontend/components/pagepile-tool.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { PagePileResponse } from "../types/search-tool.type";
 import toast from "react-hot-toast";
 import LoadingOval from "./loading-oval.component";
-import ArticlesTable from "./articles-table";
+import ArticlesTable from "./articles-table.component";
 
 export default function PagePileTool() {
   const [pagePileID, setPagePileID] = useState<string>("");
