+++ app/frontend/components/user-set-tool.tsx
@@ -3,7 +3,7 @@ import Select from "react-select";
 import toast from "react-hot-toast";
 import { UserSetResponse } from "../types/search-tool.type";
 import LoadingOval from "./loading-oval.component";
-import ArticlesTable from "./articles-table";
+import ArticlesTable from "./articles-table.component";
 
 const userSetOptions = [
   { value: "", label: "-- Choose an option --" },
